+++ pyppeteer/network_manager.py
@@ -65,8 +65,9 @@ class NetworkManager(EventEmitter):
         self._extraHTTPHeaders = OrderedDict()
         for k, v in extraHTTPHeaders.items():
             if not isinstance(v, str):
-                raise TypeError(f'Expected value of header {k} to be string, '
-                                'but {} is found.'.format(type(v)))
+                raise TypeError(
+                    f'Expected value of header "{k}" to be string, '
+                    'but {} is found.'.format(type(v)))
             self._extraHTTPHeaders[k.lower()] = v
         await self._client.send('Network.setExtraHTTPHeaders',
                                 {'headers': self._extraHTTPHeaders})
@@ -305,8 +306,8 @@ class Request(object):
 
         self._url = url
         self._resourceType = resourceType.lower()
-        self._method = payload.get('method', '')
-        self._postData = payload.get('postData', '')
+        self._method = payload.get('method')
+        self._postData = payload.get('postData')
         headers = payload.get('headers', {})
         self._headers = {k.lower(): v for k, v in headers.items()}
         self._frame = frame
@@ -331,12 +332,12 @@ class Request(object):
         return self._resourceType
 
     @property
-    def method(self) -> str:
+    def method(self) -> Optional[str]:
         """Return this request's method (GET, POST, etc.)."""
         return self._method
 
     @property
-    def postData(self) -> str:
+    def postData(self) -> Optional[str]:
         """Return post body of this request."""
         return self._postData